+++ configure
@@ -582,7 +582,7 @@ PACKAGE_STRING=
 PACKAGE_BUGREPORT=
 PACKAGE_URL=
 
-ac_unique_file="src/00headers_dirigible.cpp"
+ac_unique_file="src/00_gdal.cpp"
 # Factoring default headers for most tests.
 ac_includes_default="\
 #include <stdio.h>
